[BUGFIX][EWS-2807] Fixed fractional part

+++ presentation/src/main/java/com/elta/android/presentation/features/calcutator/viewmodel/DishDetailViewModel.kt
@@ -133,7 +133,7 @@ class DishDetailViewModel @Inject constructor(
                             setText(newDish.servings.findOrFirst { saveServing.id == it.id }.servingDescription)
                         }
                     }
-                    portionCountTextField.setText(newDish.servingAmount.toInt().toString())
+                    portionCountTextField.setText(newDish.servingAmount.toString())
                 }
         }
     }
